[FEAT] Rappi order to Open Delivery Order

+++ Src/BizPik.Orders.Hub/Modules/Core/Orders/Domain/Entity/Payment/OrderPayment.cs
@@ -4,6 +4,6 @@ namespace BizPik.Orders.Hub.Modules.Core.Orders.Domain.Entity.Payment;
 
 public record OrderPayment(
     [property: JsonPropertyName("prepaid")] int Prepaid,
-    [property: JsonPropertyName("pending")] double Pending,
+    [property: JsonPropertyName("pending")] decimal Pending,
     [property: JsonPropertyName("methods")] IReadOnlyList<OrderPaymentMethod> Methods
 );
